// lib/api/model/modelSchedule.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Jeffries Homeapp API
 * Backend REST API for Jeffries Homeapp.
 * OpenAPI spec version: 1.0
 */

export interface ModelSchedule {
  beschrijving?: string;
  dag?: string;
  duur?: number;
  eind_datum?: string;
  eind_tijd?: string;
  event_id?: string;
  heledag?: boolean;
  id?: string;
  locatie?: string;
  prioriteit?: number;
  shift_type?: string;
  start_datum?: string;
  start_tijd?: string;
  status?: string;
  team?: string;
  titel?: string;
  user_id?: string;
  weeknr?: string;
  werktijd?: string;
}