// lib/api/model/modelScheduleMeta.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Jeffries Homeapp API
 * Backend REST API for Jeffries Homeapp.
 * OpenAPI spec version: 1.0
 */

export interface ModelScheduleMeta {
  file_name?: string;
  id?: string;
  imported_at?: string;
  total_rows?: number;
  user_id?: string;
}